Se agrego la sesion

+++ templates/movies.tpl
@@ -9,11 +9,13 @@
     <p class="card-text">{{$movie->descripcion|truncate:100}}</p>
     <div class="btn-container">
       <a href="item/{{$movie->ID}}" class="btn btn-primary">Ver Mas</a>
-      <a href="delete/{{$movie->ID}}" class="btn btn-danger">Eliminar</a>
-      <a href="showEditMovie/{{$movie->ID}}" class="btn btn-warning">Editar</a>
+      {{if isset($smarty.session.ID_USER)}}
+        <a href="delete/{{$movie->ID}}" class="btn btn-danger">Eliminar</a>
+        <a href="showEditMovie/{{$movie->ID}}" class="btn btn-warning">Editar</a>
+      {{/if}}
     </div>
   </div>
-    <div class="">
+    <div>
     <p>Genero: {{$movie->nombre}}</p>
   </div>
   </div>
